pyproject.toml
[project]
name = "insider-trade-detection"
version = "0.1.0"
description = "Polymarket insider trading detection using on-chain behavioral analysis"
requires-python = ">=3.12"
dependencies = [
    "web3>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "rich>=13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
